refactor user preferences and profile page
