+++ app/javascript/src/config/constants/strings.js
