+++ app/javascript/src/config/routes.js
@@ -1,6 +1,5 @@
 export const ROOT_PATH = "/";
 
-
 // Settings page
 export const SETTINGS_TEAMS_ROUTE = "/settings/teams";
 export const SETTINGS_TEAM_ROUTE = "/settings/teams/:id";
