+++ app/javascript/src/scenes/SettingsPage/scenes/preferences/components/NotificationsGroup.jsx
