+++ app/javascript/src/scenes/SettingsPage/scenes/preferences/components/NotificationsSwitch.jsx
