+++ app/javascript/src/scenes/SettingsPage/scenes/preferences/components/NotificationsSwitchGroup.jsx
@@ -8,7 +8,7 @@ import {
   ASSIGNMENT_NOTIFICATION,
   RECENT_NOTIFICATION,
   SYSTEM_NOTIFICATION
-} from "../constants";
+} from "../../../../../config/constants/strings";
 
 import {
   changeAssignmentsNotification,
