+++ app/javascript/src/scenes/SettingsPage/scenes/profile/components/Avatar.jsx
