+++ app/javascript/src/scenes/SettingsPage/scenes/profile/components/MyStatisticsBox.jsx
