+++ app/javascript/src/scenes/SettingsPage/scenes/profile/index.jsx
@@ -1,16 +1,16 @@
 import React from "react";
-
-import MyProfile from "./MyProfile";
-import MyStatistics from "./MyStatistics";
+import SettingsAccountWrapper from "../../components/SettingsAccountWrapper";
+import MyProfile from "./components/MyProfile";
+import MyStatistics from "./components/MyStatistics";
 
 const SettingsProfile = () =>
-  <div>
+  <SettingsAccountWrapper>
     <div className="col-xs-12 col-sm-4">
       <MyProfile />
     </div>
     <div className="col-xs-12 col-sm-5">
       <MyStatistics />
     </div>
-  </div>;
+  </SettingsAccountWrapper>;
 
 export default SettingsProfile;
