+++ app/javascript/src/services/api/config.js
@@ -1,0 +1,8 @@
+import axios from "axios";
+
+export const axiosInstance = axios.create({
+  withCredentials: true,
+  headers: {
+    "X-CSRF-TOKEN": document.querySelector('meta[name="csrf-token"]').content
+  }
+});
