+++ app/javascript/src/services/api/endpoints.js
@@ -17,6 +17,7 @@ export const SEARCH_PATH = "/search";
 export const RECENT_NOTIFICATIONS_PATH = "/client_api/recent_notifications";
 
 // users
+export const USER_PROFILE_INFO = "/client_api/users/profile_info";
 
 // info dropdown_title
 export const CUSTOMER_SUPPORT_LINK = "http://scinote.net/support";
